Using double for cost values

+++ include/Planners/PathGenerator.hpp
@@ -106,9 +106,7 @@ namespace Planners
          * @return true 
          * @return false 
          */
-        bool configureCellCost(const Vec3i &coordinates_, const unsigned int &_cost);
-        // JAC: Precision
-        // bool configureCellCost(const Vec3i &coordinates_, float &_cost);
+        bool configureCellCost(const Vec3i &coordinates_, const double &_cost);
 
         /**
          * @brief Check if a set of discrete coordinates are marked as occupied
